Unify FIR and IR diagnostic tests
