+++ compiler-tests/src/test/data/diagnostic/ir/MissingMemberInjectionShouldFailBinding1.fir.ir.diag.txt
@@ -1,8 +1,7 @@
-/MissingMemberInjectionShouldFailBinding1.kt:(638,662): error: [Metro/MissingBinding] Cannot find an @Inject constructor or @Provides-annotated function/property for: Implementation
+/MissingMemberInjectionShouldFailBinding1.kt:(702,726): error: [Metro/MissingBinding] Cannot find an @Inject constructor or @Provides-annotated function/property for: Implementation
 
     Implementation is injected at
         [TestGraph.$$ContributedSubscopeGraph] SubscopeGraph#bindImplementation(…, instance)
 
 Similar bindings:
-  - Base (Supertype). Type: Alias. Source: MissingMemberInjectionShouldFailBinding1.kt:26:3 
-
+  - Base (Supertype). Type: Alias. Source: MissingMemberInjectionShouldFailBinding1.kt:28:3
