+++ compiler-tests/src/test/data/diagnostic/ir/MissingMemberInjectionShouldFailBinding2.fir.ir.diag.txt
@@ -1,5 +1,4 @@
-/MissingMemberInjectionShouldFailBinding2.kt:(441,495): error: [Metro/MissingBinding] Cannot find an @Inject constructor or @Provides-annotated function/property for: dev.zacsweers.metro.MembersInjector<ClassWithoutMembersInjector>
+/MissingMemberInjectionShouldFailBinding2.kt:(505,559): error: [Metro/MissingBinding] Cannot find an @Inject constructor or @Provides-annotated function/property for: dev.zacsweers.metro.MembersInjector<ClassWithoutMembersInjector>
 
     dev.zacsweers.metro.MembersInjector<ClassWithoutMembersInjector> is injected at
         [TestGraph.$$ContributedSubscopeGraph] SubscopeMultibindingModule#bindClassWithoutMembersInjector(…, instance)
-
