+++ compiler-tests/src/test/data/diagnostic/ir/MissingMemberInjectionShouldFailBinding2.kt
@@ -1,3 +1,5 @@
+// RUN_PIPELINE_TILL: FIR2IR
+// RENDER_IR_DIAGNOSTICS_FULL_TEXT
 // Repro for https://github.com/ZacSweers/metro/issues/656
 import kotlin.reflect.KClass
