+++ compiler-tests/src/test/data/diagnostic/ir/cycles/CyclesShouldFailAcrossMultipleGraphs.fir.ir.diag.txt
@@ -1,11 +1,11 @@
-/CyclesShouldFailAcrossMultipleGraphs.kt:(70,87): error: [Metro/GraphDependencyCycle] Graph dependency cycle detected!
+/CyclesShouldFailAcrossMultipleGraphs.kt:(134,151): error: [Metro/GraphDependencyCycle] Graph dependency cycle detected!
     test.StringGraph is requested at
         [test.CharSequenceGraph] test.StringGraph.Factory#create()
     test.CharSequenceGraph is requested at
         [test.CharSequenceGraph] test.CharSequenceGraph.Factory#create()
 
 
-/CyclesShouldFailAcrossMultipleGraphs.kt:(348,359): error: [Metro/GraphDependencyCycle] Graph dependency cycle detected!
+/CyclesShouldFailAcrossMultipleGraphs.kt:(412,423): error: [Metro/GraphDependencyCycle] Graph dependency cycle detected!
     test.CharSequenceGraph is requested at
         [test.StringGraph] test.CharSequenceGraph.Factory#create()
     test.StringGraph is requested at
