+++ compiler-tests/src/test/data/diagnostic/ir/cycles/CyclesShouldFailAcrossMultipleGraphs.kt
@@ -1,3 +1,5 @@
+// RUN_PIPELINE_TILL: FIR2IR
+// RENDER_IR_DIAGNOSTICS_FULL_TEXT
 package test
 
 @DependencyGraph
