+++ compiler-tests/src/test/data/dump/fir/aggregation/ContributingTypesDependency.fir.txt
@@ -7,7 +7,7 @@ FILE: module_lib_ContributingTypesDependency.kt
     }
 Module: main
 FILE: module_main_ContributingTypesDependency.kt
-    @R|dev/zacsweers/metro/DependencyGraph|(scope = <getClass>(Q|dev/zacsweers/metro/AppScope|)) public abstract interface ExampleGraph : R|kotlin/Any| {
+    @R|dev/zacsweers/metro/DependencyGraph|(scope = <getClass>(Q|dev/zacsweers/metro/AppScope|)) public abstract interface ExampleGraph : R|kotlin/Any|, R|ContributedInterface.$$MetroContributionToAppScope| {
         @R|kotlin/Deprecated|(message = String(This synthesized declaration should not be used directly), level = Q|kotlin/DeprecationLevel|.R|kotlin/DeprecationLevel.HIDDEN|) public final class $$MetroGraph : R|ExampleGraph| {
             private constructor(): R|ExampleGraph.$$MetroGraph| {
                 super<R|kotlin/Any|>()
