+++ compiler-tests/src/test/kotlin/dev/zacsweers/metro/compiler/AbstractFirDumpTest.kt
@@ -2,29 +2,18 @@
 // SPDX-License-Identifier: Apache-2.0
 package dev.zacsweers.metro.compiler
 
-import org.jetbrains.kotlin.config.JvmTarget
 import org.jetbrains.kotlin.test.builders.TestConfigurationBuilder
 import org.jetbrains.kotlin.test.directives.FirDiagnosticsDirectives.FIR_DUMP
-import org.jetbrains.kotlin.test.directives.JvmEnvironmentConfigurationDirectives.FULL_JDK
-import org.jetbrains.kotlin.test.directives.JvmEnvironmentConfigurationDirectives.JVM_TARGET
-import org.jetbrains.kotlin.test.runners.AbstractFirLightTreeDiagnosticsTest
-import org.jetbrains.kotlin.test.services.KotlinStandardLibrariesPathProvider
+import org.jetbrains.kotlin.test.directives.TestPhaseDirectives.RUN_PIPELINE_TILL
+import org.jetbrains.kotlin.test.services.TestPhase
 
-open class AbstractFirDumpTest : AbstractFirLightTreeDiagnosticsTest() {
-  override fun createKotlinStandardLibrariesPathProvider(): KotlinStandardLibrariesPathProvider {
-    return ClasspathBasedStandardLibrariesPathProvider
-  }
-
+open class AbstractFirDumpTest : AbstractDiagnosticTest() {
   override fun configure(builder: TestConfigurationBuilder) {
     super.configure(builder)
 
     with(builder) {
-      configurePlugin()
-
       defaultDirectives {
-        JVM_TARGET.with(JvmTarget.JVM_11)
-        +FULL_JDK
-
+        RUN_PIPELINE_TILL.with(TestPhase.FRONTEND)
         +FIR_DUMP
       }
     }
